fixed gilded altar plugin
