+++ runelite-client/src/main/java/net/runelite/client/plugins/microbot/prayer/GildedAltarOverlay.java
@@ -1,24 +1,19 @@
 package net.runelite.client.plugins.microbot.prayer;
 
-import java.awt.Color;
-import java.awt.Dimension;
-import java.awt.Graphics2D;
-
-import javax.inject.Inject;
-
 import net.runelite.client.ui.overlay.OverlayPanel;
 import net.runelite.client.ui.overlay.OverlayPosition;
 import net.runelite.client.ui.overlay.components.LineComponent;
 import net.runelite.client.ui.overlay.components.TitleComponent;
 
+import javax.inject.Inject;
+import java.awt.*;
+
 public class GildedAltarOverlay extends OverlayPanel {
-    private final GildedAltarPlugin plugin;
 
     @Inject
-    GildedAltarOverlay(GildedAltarPlugin plugin)
+    GildedAltarOverlay()
     {
-        super(plugin);
-        this.plugin = plugin;
+        super();
         setPosition(OverlayPosition.TOP_LEFT);
         setNaughty();
     }
@@ -27,14 +22,14 @@ public class GildedAltarOverlay extends OverlayPanel {
         try {
             panelComponent.setPreferredSize(new Dimension(200, 300));
             panelComponent.getChildren().add(TitleComponent.builder()
-                    .text("Micro Gilded Altar V1.0")
+                    .text("Micro Gilded Altar V1.1.0")
                     .color(Color.GREEN)
                     .build());
 
             panelComponent.getChildren().add(LineComponent.builder().build());
 
             panelComponent.getChildren().add(LineComponent.builder()
-                    .left(plugin.state.toString())
+                    .left(GildedAltarScript.state.toString())
                     .build());
 
         } catch(Exception ex) {
